backend/requirements.txt
fastapi==0.98
openpyxl
passlib
pydantic==1.10.9
python-jose
python-multipart
SQLAlchemy==2.0.17
tqdm
uvicorn[standard]